+++ app/src/main/AndroidManifest.xml
@@ -24,7 +24,8 @@
             android:name=".views.MainActivity"
             android:configChanges="keyboardHidden|screenSize|screenLayout"
             android:label="@string/app_name"
-            android:theme="@style/AppTheme.NoActionBar">
+            android:theme="@style/AppTheme.NoActionBar"
+            android:launchMode="singleInstance">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
