+++ app/src/main/java/com/testerhome/nativeandroid/views/UserProfileActivity.java
@@ -1,12 +1,18 @@
 package com.testerhome.nativeandroid.views;
 
 import android.content.Intent;
+import android.net.Uri;
 import android.os.Bundle;
+import android.widget.TextView;
 
+import com.facebook.drawee.view.SimpleDraweeView;
+import com.testerhome.nativeandroid.Config;
 import com.testerhome.nativeandroid.R;
 import com.testerhome.nativeandroid.auth.TesterHomeAccountService;
+import com.testerhome.nativeandroid.models.TesterUser;
 import com.testerhome.nativeandroid.views.base.BackBaseActivity;
 
+import butterknife.Bind;
 import butterknife.OnClick;
 
 /**
@@ -14,26 +20,54 @@ import butterknife.OnClick;
  */
 public class UserProfileActivity extends BackBaseActivity {
 
+    @Bind(R.id.id_user_avatar)
+    SimpleDraweeView userAvatar;
+    @Bind(R.id.loginName)
+    TextView userName;
+    @Bind(R.id.id_company)
+    TextView commanyName;
+    @Bind(R.id.id_tag_line)
+    TextView tagLine;
+
+    private TesterUser mTesterHomeAccount;
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_user_profile);
 
         setCustomTitle("个人资料");
+
+        mTesterHomeAccount = TesterHomeAccountService.getInstance(this).getActiveAccountInfo();
+        setupView();
     }
 
-    @OnClick(R.id.btn_show_favorite)
-    void onFavoriteClick(){
+    private void setupView() {
+
+        userAvatar.setImageURI(Uri.parse(Config.getImageUrl(mTesterHomeAccount.getAvatar_url())));
+        userName.setText(mTesterHomeAccount.getLogin());
+        commanyName.setText(mTesterHomeAccount.getCompany());
+        tagLine.setText(mTesterHomeAccount.getTagline());
+    }
+
+    @OnClick(R.id.id_create_layout)
+    void onTopicsClick() {
+        startActivity(new Intent(this, AccountTopicsActivity.class));
+    }
+
+    @OnClick(R.id.id_collect_layout)
+    void onFavoriteClick() {
         startActivity(new Intent(this, AccountFavoriteActivity.class));
     }
 
-    @OnClick(R.id.btn_show_topics)
-    void onTopicsClick(){
-        startActivity(new Intent(this, AccountTopicsActivity.class));
+    @OnClick(R.id.id_notification_layout)
+    void onNotificationClick() {
+        startActivity(new Intent(this, AccountNotificationActivity.class));
     }
 
     @OnClick(R.id.btn_logout)
     void onLogoutClick() {
         TesterHomeAccountService.getInstance(this).logout();
+        this.finish();
     }
 }
