+++ app/src/main/res/drawable-v21/item_selector.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <item android:drawable="?attr/colorPrimaryDark" android:state_pressed="true" />
+    <item android:drawable="@drawable/item_selected_drawable" android:state_pressed="true" />
     <item android:drawable="@color/white" />
 </selector>
