+++ app/src/main/res/layout/list_item_topic.xml
@@ -3,12 +3,13 @@
     xmlns:fresco="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/rl_topic_item"
+    android:background="@drawable/item_selector"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:layout_marginBottom="8dp"
-    android:layout_marginLeft="16dp"
-    android:layout_marginRight="16dp"
-    android:layout_marginTop="8dp">
+    android:paddingBottom="8dp"
+    android:paddingLeft="16dp"
+    android:paddingRight="16dp"
+    android:paddingTop="8dp">
 
     <com.facebook.drawee.view.SimpleDraweeView
         android:id="@+id/sdv_topic_user_avatar"
